BUG/MINOR: prevent unconditional override of http-server-close annotation
You cannot use the `http-server-close` annotation, as it will always be
ignored by the controller in favor of the `http-keep-alive` annotation,
even if `http-keep-alive` is empty, undefined, or false.

Both the `http-keep-alive` and `http-server-close` annotations change
the value of `defaults.HTTPConnectionMode`. But since the
`http-keep-alive` annotation is evaluated last, it overwrites whatever
value `http-server-close` has previously set.

In this commit, the value of `defaults.HTTPConnectionMode` will only
be changed if its value matches the evaluated annotation.